fix bug of redudant items
