+++ src/routes.js
@@ -1,5 +1,5 @@
 const Router = require('express');
-const products = require('./controladores/produtos');
+const productsController = require('./controllers/productsController');
 
 const route = Router();
 
@@ -7,6 +7,6 @@ route.get('/health', (req, res) => {
   return res.json({ message: 'Health Check!' });
 });
 
-route.get('/products', products.listProducts);
+route.get('/products', productsController.listProducts);
 
 module.exports = { route };
